+++ hooks/src/lib/hooks/CardPassword/utils/index.ts
@@ -1,4 +1,4 @@
-import { ErrorState, ValidationRule } from "@lib/types";
+import { ErrorState, ValidationRule } from "@/types";
 import { commonConditions, getErrorByRules } from "@utils/validation";
 
 const cardPasswordValidationRules: ValidationRule<string>[] = [
