--- hooks/src/types.ts
@@ -1,4 +1,0 @@
-export interface ValidationResult {
-  isValid: boolean;
-  errorMessage: string;
-}
