+++ hooks/src/types/index.ts
@@ -11,3 +11,8 @@ export type ValidationRule<T> = {
   condition: (value: T) => boolean;
   errorMessage: string;
 };
+
+export interface ValidationResult {
+  isValid: boolean;
+  errorMessage: string;
+}
